pyproject.toml
[project]
name = "test-that"
version = "0.1.0"
description = "A Python testing library that tells you what failed, not makes you guess"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["testing", "test", "unittest", "assertion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "jsonschema>=4.23.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
pycharm = [
    "pytest>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/dawsons-creek/test-that"
Repository = "https://github.com/dawsons-creek/test-that"
Issues = "https://github.com/dawsons-creek/test-that/issues"

[project.scripts]
that = "test_that.__main__:main"
that-pycharm = "test_that.ide.pycharm:main"

[project.entry-points."test_that.plugins"]
replay = "test_that.plugins.replay:ReplayPlugin"
json_schema = "test_that.plugins.json_schema:JSONSchemaPlugin"
lifecycle = "test_that.plugins.lifecycle:ExampleLifecyclePlugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/test_that"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
    "pymdown-extensions>=10.0",
]

[tool.that]
test_dir = "tests"
pattern = "test_*.py"
verbose = false
color = true

# Plugin system configuration
[tool.that.plugins]
# Enable/disable plugins (empty list means all available plugins are enabled)
enabled = []
disabled = []
auto_discover = true
fail_on_plugin_error = false

# Plugin-specific configuration
[tool.that.plugins.replay]
recordings_dir = "tests/recordings"
default_http_mode = "once"
time_format = "iso"
http_timeout = 30

# JSON Schema plugin (enabled by default)
[tool.that.plugins.json_schema]
# Prefer jsonschema library if available, fall back to built-in validation
prefer_jsonschema_library = true

# Lifecycle plugin (disabled by default)
[tool.that.plugins.lifecycle]
enabled = false
verbose = false
slow_threshold = 0.1
